refactor(recipes): extract page composition sections
