+++ src/features/recipes/components/RecipesPage.tsx
@@ -2,8 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 
 import { recipeListQueryOptions } from "../queries/recipeQueryOptions";
 
-import { RecipePreviewCard } from "./RecipePreviewCard";
-import { RecipesEmptyState } from "./RecipesEmptyState";
+import { RecipesPageContent } from "./RecipesPageContent";
 import { RecipesPageHeader } from "./RecipesPageHeader";
 import { RecipesPageLoading } from "./RecipesPageLoading";
 
@@ -21,15 +20,7 @@ export function RecipesPage(): JSX.Element {
   return (
     <main className="mx-auto flex max-w-6xl flex-col gap-6 py-6">
       <RecipesPageHeader recipeCount={recipes.length} />
-      {recipes.length === 0 ? (
-        <RecipesEmptyState />
-      ) : (
-        <section className="grid gap-4">
-          {recipes.map((recipe) => (
-            <RecipePreviewCard key={recipe.id} recipe={recipe} />
-          ))}
-        </section>
-      )}
+      <RecipesPageContent recipes={recipes} />
     </main>
   );
 }
